+++ app/src/main/java/com/gb/stopwatch/model/time/ElapsedTimeCalculator.kt
@@ -1,0 +1,14 @@
+package com.gb.stopwatch.model.time
+
+class ElapsedTimeCalculator(private val timestampProvider: TimestampProvider) {
+
+    fun calculate(state: StopwatchState.Running): Long {
+        val currentTimestamp = timestampProvider.getMilliseconds()
+        val timePassedSinceStart = if (currentTimestamp > state.startTime) {
+            currentTimestamp - state.startTime
+        } else {
+            0
+        }
+        return timePassedSinceStart + state.elapsedTime
+    }
+}
