+++ app/src/main/java/com/gb/stopwatch/model/time/StopwatchState.kt
@@ -1,0 +1,8 @@
+package com.gb.stopwatch.model.time
+
+sealed class StopwatchState {
+
+    data class Paused(val elapsedTime: Long) : StopwatchState()
+
+    data class Running(val startTime: Long, val elapsedTime: Long) : StopwatchState()
+}
