+++ app/src/main/java/com/gb/stopwatch/model/time/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch.model.time
+
+interface TimestampProvider {
+    fun getMilliseconds() : Long
+}
