+++ app/src/main/java/com/gb/stopwatch/viewmodel/StopwatchListOrchestrator.kt
@@ -1,0 +1,52 @@
+package com.gb.stopwatch.viewmodel
+
+import com.gb.stopwatch.model.time.StopwatchStateHolder
+import com.gb.stopwatch.model.time.TimestampMillisecondsFormatter
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class StopwatchListOrchestrator(
+    private val stopwatchStateHolder: StopwatchStateHolder,
+    private val scope: CoroutineScope
+) {
+    private var job: Job? = null
+    private val mutableTicker = MutableStateFlow(TimestampMillisecondsFormatter.DEFAULT_TIME)
+    val ticker: StateFlow<String> = mutableTicker
+
+    fun start() {
+        if (job == null) startJob()
+        stopwatchStateHolder.start()
+    }
+
+    private fun startJob() {
+        scope.launch {
+            while (isActive) {
+                mutableTicker.value = stopwatchStateHolder.getStringTimeRepresentation()
+                delay(20)
+            }
+        }
+    }
+
+    fun pause() {
+        stopwatchStateHolder.pause()
+        stopJob()
+    }
+
+    fun stop() {
+        stopwatchStateHolder.stop()
+        stopJob()
+        clearValue()
+    }
+
+    private fun clearValue() {
+        mutableTicker.value = TimestampMillisecondsFormatter.DEFAULT_TIME
+    }
+
+    private fun stopJob() {
+        scope.coroutineContext.cancelChildren()
+        job = null
+    }
+
+
+}
